v1.0.0 Release (Testnet-V3) 🚀

See merge request pulsechaincom/lighthouse-pulse!1

+++ consensus/proto_array/src/fork_choice_test_definition.rs
@@ -61,7 +61,7 @@ pub enum Operation {
     },
     AssertWeight {
         block_root: Hash256,
-        weight: u64,
+        weight: u128,
     },
 }
